requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
psycopg2==2.9.9
pillow==10.2.0
drf-nested-routers==0.93.5
